events and venues crudcompleted
