+++ src/main/kotlin/com/ticketera/TicketeraApplication.kt
@@ -1,0 +1,11 @@
+package com.ticketera
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class TicketeraApplication
+
+fun main(args: Array<String>) {
+	runApplication<TicketeraApplication>(*args)
+}
