+++ src/main/kotlin/com/ticketera/exceptions/TicketeraException.kt
@@ -1,0 +1,4 @@
+package com.ticketera.exceptions
+
+
+data class TicketeraException(val error: ErrorMessage) : Exception()
