+++ src/main/kotlin/com/ticketera/model/User.kt
@@ -1,0 +1,8 @@
+package com.ticketera.model
+
+data class User(
+    val name: String,
+    val familyName: String,
+    val email: String,
+    val roles: Set<UserRole>
+)
