+++ src/main/kotlin/com/ticketera/model/UserRole.kt
@@ -1,0 +1,5 @@
+package com.ticketera.model
+
+enum class UserRole {
+    ADMIN, USER, ORGANIZER
+}
