+++ src/main/kotlin/com/ticketera/repositories/EventRepository.kt
@@ -1,0 +1,9 @@
+package com.ticketera.repositories
+
+import com.ticketera.model.Event
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+import java.util.UUID
+
+@Repository
+interface EventRepository : CrudRepository<Event, UUID>
