+++ src/main/kotlin/com/ticketera/repositories/VenueRepository.kt
@@ -1,0 +1,9 @@
+package com.ticketera.repositories
+
+import com.ticketera.model.Venue
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+import java.util.UUID
+
+@Repository
+interface VenueRepository : CrudRepository<Venue, UUID>
