--- src/main/kotlin/com/ticketera/ticketera/TicketeraApplication.kt
@@ -1,11 +1,0 @@
-package com.ticketera.ticketera
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class TicketeraApplication
-
-fun main(args: Array<String>) {
-	runApplication<TicketeraApplication>(*args)
-}
