--- src/main/kotlin/com/ticketera/ticketera/exceptions/TicketeraException.kt
@@ -1,4 +1,0 @@
-package com.ticketera.ticketera.exceptions
-
-
-data class TicketeraException(val error: ErrorMessage) : Exception()
