--- src/main/kotlin/com/ticketera/ticketera/repositories/VenueRepository.kt
@@ -1,9 +1,0 @@
-package com.ticketera.ticketera.repositories
-
-import com.ticketera.ticketera.model.Venue
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-import java.util.UUID
-
-@Repository
-interface VenueRepository : CrudRepository<Venue, UUID>
