--- src/test/kotlin/com/ticketera/ticketera/integrationtests/TicketeraApplicationTests.kt
@@ -1,42 +1,0 @@
-package com.ticketera.ticketera.integrationtests
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.DynamicPropertyRegistry
-import org.springframework.test.context.DynamicPropertySource
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
-import org.testcontainers.junit.jupiter.Testcontainers
-
-@SpringBootTest
-@Testcontainers
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class TicketeraApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-    companion object {
-        @Container
-        val postgres = PostgreSQLContainer<Nothing>("postgres:16.4")
-            .apply {
-                withDatabaseName("ticketera")
-                withUsername("ticketera")
-                withPassword("dGVzdGRi")
-                withReuse(true)
-            }
-
-        @JvmStatic
-        @DynamicPropertySource
-        fun registerProperties(registry: DynamicPropertyRegistry) {
-            registry.add("spring.datasource.url", postgres::getJdbcUrl)
-            registry.add("spring.datasource.username", postgres::getUsername)
-            registry.add("spring.datasource.password", postgres::getPassword)
-            registry.add("spring.flyway.url", postgres::getJdbcUrl)
-            registry.add("spring.flyway.user", postgres::getUsername)
-            registry.add("spring.flyway.password", postgres::getPassword)
-        }
-    }
-}
